pyproject.toml
[tool.poetry]
name = "first_app"
version = "0.1.0"
description = ""
authors = ["Younes Piro <piroyounes148@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
Flask = "^2.0.3"
pymongo = "^4.0.0"
dnspython = "^2.2.0"
python-dotenv = "^0.19.2"
Flask-GraphQL = "^2.0.1"
graphene = "2.1.7"
mongoengine = "^0.24.0"
graphene-mongo = "0.1.17"
mongoengine-migrate = "^0.0.2"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
